requirements.txt
flask
openai-whisper
TTS
requests
werkzeug
torch
torchaudio
ffmpeg-python